[REL] stock_picking_control to V11. [FIX] Separate view and py files by model. [FIX] Update README.
