+++ stock_picking_control/README.rst
@@ -1,53 +1,49 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-   :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+.. |company| replace:: ADHOC SA
+
+.. |company_logo| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-logo.png
+   :alt: ADHOC SA
+   :target: https://www.adhoc.com.ar
+
+.. |icon| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-icon.png
+
+.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+   :target: https://www.gnu.org/licenses/agpl
    :alt: License: AGPL-3
 
-=====================
-Stock Picking Control
-=====================
+=================
+Location Security
+=================
 
 This module add this optional constraints configurable by Picking Type:
 * Block Picking Edit: Restrict add lines, change parters and other fields edition on pickings of this type. This will only apply to users with group Restrict Edit Blocked Pickings. It also block pickings duplicate.
 * Block additional quantiy: Block in pack operation line to send more quantity for products than has in the sale order.
 * Block picking deletion: Do not allow to remove pickings.
 
-
 Installation
 ============
 
 To install this module, you need to:
 
-#. Just install this module.
-
+#. Only need to install the module
 
 Configuration
 =============
 
 To configure this module, you need to:
 
-#. Set your options on each picking type
-
+#. Go to Settings/Users/User select an user then in Preference set "Restrict Location".
+#. In page "Allowed Stock Locations" set the locations do you allow.
 
 Usage
 =====
 
 To use this module, you need to:
 
-#. Just use this module.
-
-
+#. No usage needed.
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.adhoc.com.ar/
-
-.. repo_id is available in https://github.com/OCA/maintainer-tools/blob/master/tools/repos_with_ids.txt
-.. branch is "8.0" for example
-
-Known issues / Roadmap
-======================
-
-* ...
+   :target: http://runbot.adhoc.com.ar/
 
 Bug Tracker
 ===========
@@ -63,19 +59,16 @@ Credits
 Images
 ------
 
-* ADHOC SA: `Icon <http://fotos.subefotos.com/83fed853c1e15a8023b86b2b22d6145bo.png>`_.
+* |company| |icon|
 
 Contributors
 ------------
 
-
 Maintainer
 ----------
 
-.. image:: http://fotos.subefotos.com/83fed853c1e15a8023b86b2b22d6145bo.png
-   :alt: Odoo Community Association
-   :target: https://www.adhoc.com.ar
+|company_logo|
 
-This module is maintained by the ADHOC SA.
+This module is maintained by the |company|.
 
 To contribute to this module, please visit https://www.adhoc.com.ar.
