+++ stock_picking_control/__init__.py
