+++ stock_picking_control/models/__init__.py
@@ -2,4 +2,6 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import stock
+from . import stock_move_line
+from . import stock_picking_type
+from . import stock_picking
