+++ stock_picking_control/view/stock_picking_type_view.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<odoo>
+
+    <record id="view_picking_type_form" model="ir.ui.view">
+        <field name="name">stock.picking.type.form</field>
+        <field name="model">stock.picking.type</field>
+        <field name="inherit_id" ref="stock.view_picking_type_form"/>
+        <field name="arch" type="xml">
+            <group position="inside">
+                <group string="Stock">
+                    <field name="block_additional_quantiy"/>
+                    <field name="block_picking_deletion"/>
+                </group>
+            </group>
+        </field>
+    </record>
+
+</odoo>
