--- stock_picking_control/view/stock_view.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-
-
-        <record id="view_picking_type_form" model="ir.ui.view">
-            <field name="name">stock.picking.type.form</field>
-            <field name="model">stock.picking.type</field>
-            <field name="inherit_id" ref="stock.view_picking_type_form"/>
-            <field name="arch" type="xml">
-                <group position="inside">
-                    <group string="Stock">
-                        <field name="block_add_lines"/>
-                        <field name="block_additional_quantiy"/>
-                        <field name="block_picking_deletion"/>
-                    </group>
-                </group>
-            </field>
-        </record>
-
-
-        <record id="view_picking_form" model="ir.ui.view">
-            <field name="name">stock.picking.form</field>
-            <field name="model">stock.picking</field>
-            <field name="inherit_id" ref="stock.view_picking_form"/>
-            <!-- por si stock usability esta instalado, queremos que esta pise aquella -->
-            <field name="priority">20</field>
-            <field name="groups_id" eval="[(6, 0, [ref('group_restrict_edit_picking')])]"/>
-            <field name="arch" type="xml">
-                <field name="partner_id" position="after">
-                    <field name="block_add_lines" invisible="1"/>
-                    <field name="block_additional_quantiy" invisible="1"/>
-                </field>
-                <field name="move_lines_related" position="attributes">
-                     <!-- sacamos el done de invisible porque asi lo queremos y hacemos en stock usablity tambien -->
-                     <attribute name="attrs">{'readonly':[('block_add_lines','=',True)],'invisible': [('state', 'in', ('draft', 'confirmed', 'waiting'))]}</attribute>
-                </field>
-                <field name="move_lines" position="attributes">
-                     <attribute name="attrs">{'readonly':[('block_add_lines','=',True)],'invisible':  [('state', 'not in', ('draft', 'confirmed', 'waiting'))]}</attribute>
-                </field>
-
-                <!-- this came from tran.. personalization -->
-                <field name="partner_id" position="attributes">
-                    <attribute name="attrs">{'readonly':[('block_add_lines','=',True)]}</attribute>
-                </field>
-
-                <field name="min_date" position="attributes">
-                    <attribute name="attrs">{'readonly':[('block_add_lines','=',True)]}</attribute>
-                </field>
-                <field name="origin" position="attributes">
-                    <attribute name="attrs">{'readonly':[('block_add_lines','=',True)]}</attribute>
-                </field>
-                <field name="move_type" position="attributes">
-                    <attribute name="attrs">{'readonly':[('block_add_lines','=',True)]}</attribute>
-                </field>
-                <field name="declared_value" position="attributes">
-                    <attribute name="attrs">{'readonly':[('block_add_lines','=',True)]}</attribute>
-                </field>
-            </field>
-        </record>
-        
-    </data>
-</odoo>
