+++ stock_ux/models/__init__.py
@@ -8,3 +8,4 @@ from . import stock_return_picking
 from . import product_product
 from . import stock_warehouse_orderpoint
 from . import stock_move_line
+from . import stock_picking_type
