+++ stock_ux/models/stock_move_line.py
@@ -2,10 +2,12 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from odoo import models, api, fields
+from odoo import models, api, fields, _
+from odoo.exceptions import ValidationError
 
 
 class StockMoveLine(models.Model):
+
     _inherit = 'stock.move.line'
 
     picking_create_user_id = fields.Many2one(
@@ -27,9 +29,24 @@ class StockMoveLine(models.Model):
         readonly=True,
     )
 
+    block_additional_quantiy = fields.Boolean(
+        related='picking_id.block_additional_quantiy',
+        readonly=True,
+        default=True,
+    )
+
     @api.multi
     def set_all_done(self):
         for rec in self:
             rec.update({'qty_done': rec.move_id.product_uom_qty})
         if self._context.get('from_popup', False):
             return self[0].move_id.action_show_details()
+
+    @api.constrains('qty_done')
+    def _check_quantity(self):
+        for pack in self.filtered(
+                lambda x: x.picking_id.block_additional_quantiy
+                and x.product_qty < x.qty_done):
+            raise ValidationError(_(
+                'You can not transfer a product without a move on the '
+                'picking!'))
