+++ stock_ux/security/stock_ux_security.xml
@@ -10,4 +10,8 @@
         <field name="category_id" ref="base.module_category_hidden"/>
     </record>
 
+    <record model="res.groups" id="group_restrict_edit_picking">
+        <field name="name">Restrict Edit Blocked Pickings</field>
+    </record>
+
 </odoo>
