+++ stock_ux/views/stock_picking_view.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<odoo>
+
+    <record id="view_picking_form" model="ir.ui.view">
+        <field name="name">stock.picking.form</field>
+        <field name="model">stock.picking</field>
+        <field name="inherit_id" ref="stock.view_picking_form"/>
+
+        <!-- por si stock usability esta instalado, queremos que esta pise aquella -->
+        <field name="priority">20</field>
+        <field name="groups_id" eval="[(6, 0, [ref('group_restrict_edit_picking')])]"/>
+        <field name="arch" type="xml">
+            <field name="partner_id" position="after">
+                <field name="block_additional_quantiy" invisible="1"/>
+            </field>
+        </field>
+    </record>
+
+</odoo>
